# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Medusa Core API",
        "description": "This is the Medusa Core API. Most routes are protected with JWT authentication.",
        "contact": {
            "email": "mdoc@trmn.org"
        },
        "license": {
            "name": "GPL 3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Attempts to login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "The email for login",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "failed to login"
                    }
                }
            }
        },
        "/api/v1/branches": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Get list of branches",
                "description": "Returns a list of branches ordered by abbreviation.",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "bad request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Branches"
                ],
                "summary": "Create a new branch",
                "description": "Creates a new branch",
                "operationId": "store",
                "requestBody": {
                    "request": "Branch",
                    "description": "Branch object to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Branch"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "branch successfully created"
                    },
                    "403": {
                        "description": "permission denied"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/branches/{id}": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Retrieve the given branch",
                "description": "Retrieves the given branch",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the branch to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "branch retrieved"
                    },
                    "403": {
                        "description": "permission denied"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Branches"
                ],
                "summary": "Update the given branch",
                "description": "Updates the given branch",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the branch to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "Branch",
                    "description": "Branch object to be modified",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Branch"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "branch successfully modified"
                    },
                    "403": {
                        "description": "permission denied"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Branches"
                ],
                "summary": "Delete the given branch",
                "description": "Deletes the given branch. Note: this is a hard delete.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the branch to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "branch successfully deleted"
                    },
                    "403": {
                        "description": "permission denied"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/countries": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Get list of countries",
                "description": "Returns a list of countries by CCA3 code",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/countries/{cca3}": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Get single country data",
                "description": "Returns all data for a single country specified by CCA3 code",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "cca3",
                        "in": "path",
                        "description": "Country CCA3 code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/countries/{cca3}/states": {
            "get": {
                "tags": [
                    "Countries"
                ],
                "summary": "Get list of states",
                "description": "Returns a list of states for a given country specified by CCA3 code",
                "operationId": "getStates",
                "parameters": [
                    {
                        "name": "cca3",
                        "in": "path",
                        "description": "Country CCA3 code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns a list of users ordered by last name and first name",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "bad request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Creates a new user",
                "operationId": "store",
                "requestBody": {
                    "request": "User",
                    "description": "User object to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user successfully created"
                    },
                    "403": {
                        "description": "permission denied"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve the given user",
                "description": "Retrieves the given user, without hidden fields like password",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user retrieved"
                    },
                    "403": {
                        "description": "permission denied"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update the given user",
                "description": "Updates the given user",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "User",
                    "description": "User object to be modified",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user successfully modified"
                    },
                    "403": {
                        "description": "permission denied"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete the given user",
                "description": "Deletes the given user. Note: this is a hard delete.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user successfully deleted"
                    },
                    "403": {
                        "description": "permission denied"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Branch": {
                "properties": {
                    "abbreviation": {
                        "description": "The abbreviated version of the branch's name",
                        "type": "string"
                    },
                    "name": {
                        "description": "The full name of the branch",
                        "type": "string"
                    },
                    "is_civilian": {
                        "description": "Is this branch a civilian branch?",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "first_name": {
                        "description": "First name",
                        "type": "string"
                    },
                    "middle_name": {
                        "description": "Middle name",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Last name",
                        "type": "string"
                    },
                    "phone_number": {
                        "description": "Phone number",
                        "type": "string"
                    },
                    "date_of_birth": {
                        "description": "Date of birth, in yyyy-mm-dd format",
                        "type": "string"
                    },
                    "address_line_1": {
                        "description": "Address line one",
                        "type": "string"
                    },
                    "address_line_2": {
                        "description": "Address line two",
                        "type": "string"
                    },
                    "city": {
                        "description": "City",
                        "type": "string"
                    },
                    "province_or_state": {
                        "description": "Province or state",
                        "type": "string"
                    },
                    "postal_code": {
                        "description": "Postal code",
                        "type": "string"
                    },
                    "country": {
                        "description": "Country",
                        "type": "string"
                    },
                    "is_active": {
                        "description": "Is active?",
                        "type": "boolean"
                    },
                    "email": {
                        "description": "Email address",
                        "type": "string"
                    },
                    "password": {
                        "description": "Password",
                        "type": "string"
                    },
                    "branch_id": {
                        "description": "ID of the branch the user belongs to",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (bearer $bearertoken$)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Authentication routes."
        },
        {
            "name": "Branches",
            "description": "Branch control routes."
        },
        {
            "name": "Countries",
            "description": "Routes for standardizing country data. These are not protected by authorization."
        },
        {
            "name": "Users",
            "description": "User control routes."
        }
    ],
    "security": [
        []
    ]
}